// FRONT-END-ANGULAR/src/app/interfaces/persona.ts
export interface Persona{
    idpersona: number;
    nombres: String;
    apellido: String;
    email: String;
    ocupacion: String;
    ubicacion: String;
    username?: String;
    password?: String;
    acerca: String;
    fotoperfil: String;
    fotofondo: String;
    token?:String


}